refactor: uppercase models and modules filename
