+++ src/App.tsx
@@ -15,10 +15,10 @@ import CoreLayout from './shared/layout/CoreLayout/CoreLayout'
 
 // Code splitting
 const Home = React.lazy(() =>
-	import(/* webpackChunkName: "home" */ './pages/Home/Home'),
+	import(/* webpackChunkName: "Home" */ './pages/Home/Home'),
 )
 const About = React.lazy(() =>
-	import(/* webpackChunkName: "about" */ './pages/About/About'),
+	import(/* webpackChunkName: "About" */ './pages/About/About'),
 )
 
 // Constants
