+++ src/models/Posts.ts
@@ -1,0 +1,6 @@
+import BaseModels from './bases/BaseModels'
+import Post from './Post'
+
+export default interface Posts extends BaseModels {
+	list: [Post]
+}
