+++ src/models/bases/BaseModels.ts
@@ -1,4 +1,4 @@
-import BaseModel from './baseModel'
+import BaseModel from './BaseModel'
 
 export default interface BaseModels extends BaseModel {
 	list: any
