--- src/models/posts.ts
@@ -1,6 +1,0 @@
-import BaseModels from './bases/baseModels'
-import Post from './post'
-
-export default interface Posts extends BaseModels {
-	list: [Post]
-}
